+++ test-suite/embedded-test/filter.cpp
@@ -9,11 +9,11 @@ uint32_t printfMask = 0;
 
 class Filter : public TopicFilter {
   public:
-    void prolog([[gnu::unused]] const long topicId, [[gnu::unused]] const long len, void* data, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
+    void prolog([[gnu::unused]] const uint32_t topicId, [[gnu::unused]] const size_t len, void* data, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
         int* received = (int*)data;
         *received = *received + 1000000;
     }
-    void epilog([[gnu::unused]] const long topicId, [[gnu::unused]] const long len, [[gnu::unused]] void* data, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
+    void epilog([[gnu::unused]] const uint32_t topicId, [[gnu::unused]] const size_t len, [[gnu::unused]] void* data, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
         PRINTF("Filter changed the value\n");
     }
 } filter;
