+++ tutorials/50-time-related/1pps-getriggered-scheduling/scheduler-1pps-getriggert.cpp
@@ -65,8 +65,8 @@ void ActivityExecuter::run(void) {
 class OnePulsePerSecond : public Subscriber {
   public:
     OnePulsePerSecond() : Subscriber(interruptSigterm, "simplesub") {}
-    void putFromInterrupt([[gnu::unused]] const long topicId, [[gnu::unused]] const void* any, int len) {
-        xprintf("external Event at %3.9f datalen %d\n", SECONDS_NOW(), len);
+    void putFromInterrupt([[gnu::unused]] const uint32_t topicId, [[gnu::unused]] const void* any, size_t len) {
+        xprintf("external Event at %3.9f datalen %lu\n", SECONDS_NOW(), static_cast<unsigned long>(len));
         activityExecuter.resume();
     }
 } onePulsePerSecond;
