+++ tutorials/50-time-related/time-triegged-protocoll/server/ts-server.cpp
@@ -31,7 +31,7 @@ class TimeSyncServer : public Subscriber, public Putter {
   public:
     TimeSyncServer() : Subscriber(clientRequest, "timeSyncServer") {}
 
-    long put([[gnu::unused]] const long _topicId, [[gnu::unused]] const long _len, void* _msg, [[gnu::unused]] const NetMsgInfo& _netMsgInfo) {
+    uint32_t put([[gnu::unused]] const uint32_t _topicId, [[gnu::unused]] const size_t _len, void* _msg, [[gnu::unused]] const NetMsgInfo& _netMsgInfo) {
         int32_t requestId               = *(int32_t*)_msg;
         response.ntspTimes.rcvTS        = sysTime.getUTC(); // by real time server, set by hardware at arrival
         response.clientRequestCnt       = requestId;
@@ -46,7 +46,7 @@ class TimeSyncServer : public Subscriber, public Putter {
     }
 
     /** Just for debug: Print the utc time as response to the (global) interrupt **/
-    bool putGeneric([[gnu::unused]] const long _topicId, [[gnu::unused]] const unsigned int _msgLen, [[gnu::unused]] const void* _msg, [[gnu::unused]] const NetMsgInfo& _netMsgInfo) {
+    bool putGeneric([[gnu::unused]] const uint32_t _topicId, [[gnu::unused]] const size_t _msgLen, [[gnu::unused]] const void* _msg, [[gnu::unused]] const NetMsgInfo& _netMsgInfo) {
         PRINTF(INT_TIME_SRV "%9.6f", utcMillisecsNow());
         return true;
     }
