+++ tutorials/70-network-and-io/10-udp/udpreceiver-async.cpp
@@ -34,10 +34,10 @@ class Udpreceiver :  public Subscriber {
 public:
     Udpreceiver() : Subscriber(udpMessages, "udpReceiver") { }
 
-    void putFromInterrupt([[gnu::unused]] const long topicId, const void* data, [[gnu::unused]] int len) {
+    void putFromInterrupt([[gnu::unused]] const uint32_t topicId, const void* data, [[gnu::unused]] size_t len) {
         const GenericMsgRef* msg = (const GenericMsgRef*)data;
         char buffer[100] = {};
-        memcpy(buffer, msg->msgPtr, msg->msgLen > 100 ? 100 : msg->msgLen);
+        memcpy(buffer, msg->msgPtr, static_cast<size_t>(msg->msgLen > 100 ? 100 : msg->msgLen));
         buffer[msg->msgLen > 100? 99 : msg->msgLen - 1] = 0;
         xprintf("\n Async: %d %s\n", static_cast<int>(msg->msgLen), msg->msgPtr); // no PRINTF in interrupts (Sempahore)
     }
